+++ src/select/docs/examples/autofocus/controller.js
@@ -4,14 +4,4 @@ angular.module('selectDemo')
         $scope.shopArr = ShopArr.query();
 
         $scope.bundle = null;
-
-        $scope.bundle1 = [{
-            "id": 5,
-            "name": "shirt",
-            "category": "clothes"
-        },{
-            "id": 2,
-            "name": "shoes",
-            "category": "shoes"
-        }];
     });